git-svn-id: svn://svn.code.sf.net/p/mcapi/code@164 3a57ee49-f89d-46b9-a3c0-71fdb2aa5dce

+++ src/TestCapi/Test_DSpaceList.cpp
@@ -142,7 +142,7 @@ void DRV_DSpaceList_Alloc(UINT i)
 			|| p->pDSNode[0].pSpList->pBlock == NULL 
 			|| p->pDSNode[0].pSpList->uFreeCount != 0
 			|| (INT)pData - (INT)p->pDSNode[0].pSpList->pBlock != sizeof(SPNODE) + 127 * (sizeof(SPNODE)+32)
-			|| (INT)p->pDSNode[0].pSpList->pHead != NULL )
+			|| p->pDSNode[0].pSpList->pHead != NULL )
         {
             printf( "DSpaceList_Alloc() ≤‚ ‘”√¿˝3 ß∞‹!\n" );
         }
